Correct rewrite_actions, move overrides to own file
